swarm/src/lib: Move dial_opts to separate file
