+++ swarm/src/lib.rs
@@ -59,6 +59,7 @@ mod registry;
 mod test;
 mod upgrade;
 
+pub mod dial_opts;
 pub mod protocols_handler;
 pub mod toggle;
 
@@ -1810,202 +1811,5 @@ mod tests {
                 return Poll::Pending;
             }
         }))
-    }
-}
-
-// TODO: Move above tests or into separate module.
-pub mod dial_opts {
-    use libp2p_core::{Multiaddr, PeerId};
-
-    /// Options to configure a dial to a known or unknown peer.
-    ///
-    /// Used in [`Swarm::dial`](crate::Swarm::dial) and
-    /// [`NetworkBehaviourAction::Dial`](crate::behaviour::NetworkBehaviourAction::Dial).
-    ///
-    /// To construct use either of:
-    ///
-    /// - [`DialOpts::peer_id`] dialing a known peer
-    ///
-    /// - [`DialOpts::unknown_peer_id`] dialing an unknown peer
-    #[derive(Debug)]
-    pub struct DialOpts(pub(super) Opts);
-
-    impl DialOpts {
-        /// Dial a known peer.
-        ///
-        ///   ```
-        ///   # use libp2p_swarm::dial_opts::{DialOpts, PeerCondition};
-        ///   # use libp2p_core::PeerId;
-        ///   DialOpts::peer_id(PeerId::random())
-        ///      .condition(PeerCondition::Disconnected)
-        ///      .addresses(vec!["/ip6/::1/tcp/12345".parse().unwrap()])
-        ///      .extend_addresses_through_behaviour()
-        ///      .build();
-        ///   ```
-        pub fn peer_id(peer_id: PeerId) -> WithPeerId {
-            WithPeerId {
-                peer_id,
-                condition: Default::default(),
-            }
-        }
-
-        /// Dial an unknown peer.
-        ///
-        ///   ```
-        ///   # use libp2p_swarm::dial_opts::DialOpts;
-        ///   DialOpts::unknown_peer_id()
-        ///      .address("/ip6/::1/tcp/12345".parse().unwrap())
-        ///      .build();
-        ///   ```
-        pub fn unknown_peer_id() -> WithoutPeerId {
-            WithoutPeerId {}
-        }
-
-        /// Get the [`PeerId`] specified in a [`DialOpts`] if any.
-        pub fn get_peer_id(&self) -> Option<PeerId> {
-            match self {
-                DialOpts(Opts::WithPeerId(WithPeerId { peer_id, .. })) => Some(*peer_id),
-                DialOpts(Opts::WithPeerIdWithAddresses(WithPeerIdWithAddresses {
-                    peer_id,
-                    ..
-                })) => Some(*peer_id),
-                DialOpts(Opts::WithoutPeerIdWithAddress(_)) => None,
-            }
-        }
-    }
-
-    /// Internal options type.
-    ///
-    /// Not to be constructed manually. Use either of the below instead:
-    ///
-    /// - [`DialOpts::peer_id`] dialing a known peer
-    /// - [`DialOpts::unknown_peer_id`] dialing an unknown peer
-    #[derive(Debug)]
-    pub(super) enum Opts {
-        WithPeerId(WithPeerId),
-        WithPeerIdWithAddresses(WithPeerIdWithAddresses),
-        WithoutPeerIdWithAddress(WithoutPeerIdWithAddress),
-    }
-
-    #[derive(Debug)]
-    pub struct WithPeerId {
-        pub(crate) peer_id: PeerId,
-        pub(crate) condition: PeerCondition,
-    }
-
-    impl WithPeerId {
-        /// Specify a [`PeerCondition`] for the dial.
-        pub fn condition(mut self, condition: PeerCondition) -> Self {
-            self.condition = condition;
-            self
-        }
-
-        /// Specify a set of addresses to be used to dial the known peer.
-        pub fn addresses(self, addresses: Vec<Multiaddr>) -> WithPeerIdWithAddresses {
-            WithPeerIdWithAddresses {
-                peer_id: self.peer_id,
-                condition: self.condition,
-                addresses,
-                extend_addresses_through_behaviour: false,
-            }
-        }
-
-        /// Build the final [`DialOpts`].
-        ///
-        /// Addresses to dial the peer are retrieved via
-        /// [`NetworkBehaviour::addresses_of_peer`](crate::behaviour::NetworkBehaviour::addresses_of_peer).
-        pub fn build(self) -> DialOpts {
-            DialOpts(Opts::WithPeerId(self))
-        }
-    }
-
-    #[derive(Debug)]
-    pub struct WithPeerIdWithAddresses {
-        pub(crate) peer_id: PeerId,
-        pub(crate) condition: PeerCondition,
-        pub(crate) addresses: Vec<Multiaddr>,
-        pub(crate) extend_addresses_through_behaviour: bool,
-    }
-
-    impl WithPeerIdWithAddresses {
-        /// Specify a [`PeerCondition`] for the dial.
-        pub fn condition(mut self, condition: PeerCondition) -> Self {
-            self.condition = condition;
-            self
-        }
-
-        /// In addition to the provided addresses, extend the set via
-        /// [`NetworkBehaviour::addresses_of_peer`](crate::behaviour::NetworkBehaviour::addresses_of_peer).
-        pub fn extend_addresses_through_behaviour(mut self) -> Self {
-            self.extend_addresses_through_behaviour = true;
-            self
-        }
-
-        /// Build the final [`DialOpts`].
-        pub fn build(self) -> DialOpts {
-            DialOpts(Opts::WithPeerIdWithAddresses(self))
-        }
-    }
-
-    #[derive(Debug)]
-    pub struct WithoutPeerId {}
-
-    impl WithoutPeerId {
-        /// Specify a single address to dial the unknown peer.
-        pub fn address(self, address: Multiaddr) -> WithoutPeerIdWithAddress {
-            WithoutPeerIdWithAddress { address }
-        }
-    }
-
-    #[derive(Debug)]
-    pub struct WithoutPeerIdWithAddress {
-        pub(crate) address: Multiaddr,
-    }
-
-    impl WithoutPeerIdWithAddress {
-        /// Build the final [`DialOpts`].
-        pub fn build(self) -> DialOpts {
-            DialOpts(Opts::WithoutPeerIdWithAddress(self))
-        }
-    }
-
-    /// The available conditions under which a new dialing attempt to
-    /// a known peer is initiated.
-    ///
-    /// ```
-    /// # use libp2p_swarm::dial_opts::{DialOpts, PeerCondition};
-    /// # use libp2p_core::PeerId;
-    /// #
-    /// DialOpts::peer_id(PeerId::random())
-    ///    .condition(PeerCondition::Disconnected)
-    ///    .build();
-    /// ```
-    #[derive(Debug, Copy, Clone)]
-    pub enum PeerCondition {
-        /// A new dialing attempt is initiated _only if_ the peer is currently
-        /// considered disconnected, i.e. there is no established connection
-        /// and no ongoing dialing attempt.
-        ///
-        /// If there is an ongoing dialing attempt, the addresses reported by
-        /// [`NetworkBehaviour::addresses_of_peer`](crate::behaviour::NetworkBehaviour::addresses_of_peer)
-        /// are added to the ongoing dialing attempt, ignoring duplicates.
-        Disconnected,
-        /// A new dialing attempt is initiated _only if_ there is currently
-        /// no ongoing dialing attempt, i.e. the peer is either considered
-        /// disconnected or connected but without an ongoing dialing attempt.
-        ///
-        /// If there is an ongoing dialing attempt, the addresses reported by
-        /// [`NetworkBehaviour::addresses_of_peer`](crate::behaviour::NetworkBehaviour::addresses_of_peer)
-        /// are added to the ongoing dialing attempt, ignoring duplicates.
-        NotDialing,
-        /// A new dialing attempt is always initiated, only subject to the
-        /// configured connection limits.
-        Always,
-    }
-
-    impl Default for PeerCondition {
-        fn default() -> Self {
-            PeerCondition::Disconnected
-        }
     }
 }
